requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.4.1
    # via streamlit
annotated-types==0.6.0
    # via
    #   conva-ai
    #   pydantic
anyio==4.4.0
    # via
    #   conva-ai
    #   httpx
asyncio==3.4.3
    # via -r requirements.in
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via
    #   google-auth
    #   streamlit
certifi==2024.2.2
    # via
    #   conva-ai
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
conva-ai==0.2.4
    # via -r requirements.in
db-dtypes==1.3.1
    # via -r requirements.in
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-api-core[grpc]==2.23.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-core
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-core
google-cloud-bigquery==3.27.0
    # via -r requirements.in
google-cloud-core==2.4.1
    # via google-cloud-bigquery
google-crc32c==1.6.0
    # via google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-bigquery
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.67.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.67.1
    # via google-api-core
h11==0.14.0
    # via
    #   conva-ai
    #   httpcore
httpcore==1.0.5
    # via
    #   conva-ai
    #   httpx
httpx==0.27.0
    # via conva-ai
idna==3.7
    # via
    #   anyio
    #   conva-ai
    #   httpx
    #   requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.13.4
    # via altair
numpy==2.1.3
    # via
    #   db-dtypes
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   db-dtypes
    #   google-cloud-bigquery
    #   streamlit
pandas==2.2.3
    # via
    #   -r requirements.in
    #   db-dtypes
    #   streamlit
pillow==11.0.0
    # via streamlit
proto-plus==1.25.0
    # via google-api-core
protobuf==5.28.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==18.0.0
    # via
    #   db-dtypes
    #   streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.7.0
    # via conva-ai
pydantic-core==2.18.1
    # via
    #   conva-ai
    #   pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   conva-ai
    #   httpx
sseclient-py==1.8.0
    # via conva-ai
streamlit==1.40.1
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.1
    # via streamlit
typing-extensions==4.11.0
    # via
    #   altair
    #   conva-ai
    #   pydantic
    #   pydantic-core
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
